+++ CHANGELOG.md
@@ -13,7 +13,8 @@
 **Fixed bugs:**
 
 - ~/.gitignore_global is not handled [\#453](https://github.com/hsz/idea-gitignore/issues/#453)
-- NPE in MatcherUtil.match [\#457](https://github.com/hsz/idea-gitignore/issues/#475)
+- NPE in MatcherUtil.match [\#457](https://github.com/hsz/idea-gitignore/issues/#457)
+- 'Outer ignore rules' shows previous project's exclude files [\#460](https://github.com/hsz/idea-gitignore/issues/#460)
 
 
 ## [v2.1.1](https://github.com/hsz/idea-gitignore/tree/v2.1.1) (2017-08-23)